Update disk matching procedure
Move the disk matching procedure (by which disks containing desired
prefixes are found) into a function that is used by both
`local_transfer()` and `globus_upload()` functions. Update matching
algorithm to a list comprehension (instead of two `for` loops) using
`str.startswith()` and `str.upper()` instead of a regex match.
Add a print statement that indicates the number of cards to be
transferred, for a sanity check.